fix(desktop): unify authoritative route identity
Keep explicit connection identity authoritative only while its registry entry is live, and route every legacy compatibility inference through the same complete pre-dial auth and transport envelope used by Desktop route selection.

Absorbs the full matcher and collision matrix from #90198 into the stricter current-registration contract in #90913.
